requirements.txt
fastapi
jinja2
uvicorn
gemini
python-dotenv